pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hma"
version = "0.1.0"
description = "Human-Machine Visual Alignment benchmark skeleton."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.23",
    "pillow>=10.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
models = [
    "timm>=1.0",
]
saliency = [
    "torch>=2.0",
]

[project.scripts]
hma-dummy-pipeline = "hma.pipelines.dummy:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "--basetemp=.pytest_tmp"
